requirements.txt
Django==4.2.9
python-decouple==3.8
Pillow==10.1.0
psycopg2-binary==2.9.9
django-crispy-forms==2.1
crispy-bootstrap5==0.7
requests==2.31.0
reportlab==4.0.7
weasyprint==60.1
django-filter==23.5
django-storages==1.14.2
gunicorn==21.2.0
whitenoise==6.6.0
celery==5.3.4
redis==5.0.1
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.2.14
boto3==1.34.34
python-dotenv==1.0.0